Move bitmap classes
