+++ QuickCode/mobile/src/main/java/com/toolslab/quickcode/model/CodeFileCreator.java
@@ -11,8 +11,8 @@ import android.util.SparseArray;
 import com.google.android.gms.vision.Frame;
 import com.google.android.gms.vision.barcode.Barcode;
 import com.google.android.gms.vision.barcode.BarcodeDetector;
-import com.toolslab.quickcode.util.BitmapUtils;
 import com.toolslab.quickcode.util.TypeUtils;
+import com.toolslab.quickcode.util.bitmap.BitmapUtils;
 import com.toolslab.quickcode.util.encode.EncodingUtils;
 import com.toolslab.quickcode.util.log.Logger;
