+++ QuickCode/mobile/src/main/java/com/toolslab/quickcode/util/UriUtils.java
@@ -12,6 +12,7 @@ import android.support.annotation.CheckResult;
 import android.support.annotation.NonNull;
 import android.support.annotation.Nullable;
 
+import com.toolslab.quickcode.util.bitmap.PdfToBitmapConverter;
 import com.toolslab.quickcode.util.encode.EncodingUtils;
 import com.toolslab.quickcode.util.log.Logger;
