+++ QuickCode/mobile/src/main/java/com/toolslab/quickcode/util/bitmap/BitmapUtils.java
@@ -1,4 +1,4 @@
-package com.toolslab.quickcode.util;
+package com.toolslab.quickcode.util.bitmap;
 
 
 import android.graphics.Bitmap;
