+++ QuickCode/mobile/src/main/java/com/toolslab/quickcode/view/main/CodesListFragment.java
@@ -28,8 +28,8 @@ import com.toolslab.quickcode.model.CodeFileCreator;
 import com.toolslab.quickcode.model.CodeFileFactory;
 import com.toolslab.quickcode.model.CodeFileViewModel;
 import com.toolslab.quickcode.util.AssetPathLoader;
-import com.toolslab.quickcode.util.PdfToBitmapConverter;
 import com.toolslab.quickcode.util.UriUtils;
+import com.toolslab.quickcode.util.bitmap.PdfToBitmapConverter;
 import com.toolslab.quickcode.util.log.Tracker;
 import com.toolslab.quickcode.view.base.BaseActivity;
 import com.toolslab.quickcode.view.base.BaseFragment;
